// src/app/core/services/model/deviceTypeName.ts
/**
 * TWMS REST API
 * TWMS REST API
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export type DeviceTypeName = 'Wall-E' | 'Beat-E' | 'WireSens' | 'WirePort' | 'DTE' | 'Zirk-E' | 'Rulantica' | 'RulanticaT' | 'UfcFaultContact' | 'Hyplus' | 'Pump' | 'MagneticValve' | 'TemperatureSensor' | 'FlowSensor' | 'PressureSensor' | 'PulseSensor' | 'FlowAndTemperatureSensor' | 'GPIO' | 'LEDBar' | 'KeySwitch';

export const DeviceTypeName = {
    WallE: 'Wall-E' as DeviceTypeName,
    BeatE: 'Beat-E' as DeviceTypeName,
    WireSens: 'WireSens' as DeviceTypeName,
    WirePort: 'WirePort' as DeviceTypeName,
    DTE: 'DTE' as DeviceTypeName,
    ZirkE: 'Zirk-E' as DeviceTypeName,
    Rulantica: 'Rulantica' as DeviceTypeName,
    RulanticaT: 'RulanticaT' as DeviceTypeName,
    UfcFaultContact: 'UfcFaultContact' as DeviceTypeName,
    Hyplus: 'Hyplus' as DeviceTypeName,
    Pump: 'Pump' as DeviceTypeName,
    MagneticValve: 'MagneticValve' as DeviceTypeName,
    TemperatureSensor: 'TemperatureSensor' as DeviceTypeName,
    FlowSensor: 'FlowSensor' as DeviceTypeName,
    PressureSensor: 'PressureSensor' as DeviceTypeName,
    PulseSensor: 'PulseSensor' as DeviceTypeName,
    FlowAndTemperatureSensor: 'FlowAndTemperatureSensor' as DeviceTypeName,
    GPIO: 'GPIO' as DeviceTypeName,
    LEDBar: 'LEDBar' as DeviceTypeName,
    KeySwitch: 'KeySwitch' as DeviceTypeName
};

import { marker as _ } from '@biesbjerg/ngx-translate-extract-marker';

export const DeviceTypeNameMapping = {
  [DeviceTypeName.WallE]: _('API.DEVICE_TYPE_NAME.WallE'),
  [DeviceTypeName.BeatE]: _('API.DEVICE_TYPE_NAME.BeatE'),
  [DeviceTypeName.WireSens]: _('API.DEVICE_TYPE_NAME.WireSens'),
  [DeviceTypeName.WirePort]: _('API.DEVICE_TYPE_NAME.WirePort'),
  [DeviceTypeName.DTE]: _('API.DEVICE_TYPE_NAME.DTE'),
  [DeviceTypeName.ZirkE]: _('API.DEVICE_TYPE_NAME.ZirkE'),
  [DeviceTypeName.Rulantica]: _('API.DEVICE_TYPE_NAME.Rulantica'),
  [DeviceTypeName.RulanticaT]: _('API.DEVICE_TYPE_NAME.RulanticaT'),
  [DeviceTypeName.UfcFaultContact]: _('API.DEVICE_TYPE_NAME.UfcFaultContact'),
  [DeviceTypeName.Hyplus]: _('API.DEVICE_TYPE_NAME.Hyplus'),
  [DeviceTypeName.Pump]: _('API.DEVICE_TYPE_NAME.Pump'),
  [DeviceTypeName.MagneticValve]: _('API.DEVICE_TYPE_NAME.MagneticValve'),
  [DeviceTypeName.TemperatureSensor]: _('API.DEVICE_TYPE_NAME.TemperatureSensor'),
  [DeviceTypeName.FlowSensor]: _('API.DEVICE_TYPE_NAME.FlowSensor'),
  [DeviceTypeName.PressureSensor]: _('API.DEVICE_TYPE_NAME.PressureSensor'),
  [DeviceTypeName.PulseSensor]: _('API.DEVICE_TYPE_NAME.PulseSensor'),
  [DeviceTypeName.FlowAndTemperatureSensor]: _('API.DEVICE_TYPE_NAME.FlowAndTemperatureSensor'),
  [DeviceTypeName.GPIO]: _('API.DEVICE_TYPE_NAME.GPIO'),
  [DeviceTypeName.LEDBar]: _('API.DEVICE_TYPE_NAME.LEDBar'),
  [DeviceTypeName.KeySwitch]: _('API.DEVICE_TYPE_NAME.KeySwitch'),
};